tried to hide the mouse cursor when it is captured and discovered a bug in godot.
the mouse cursor should be hidden when captured. but it stays visible instead.
capturing and then hiding the mouse won't work, since these two modes are exclusive.
added an idea for a workaround in the comments, TODO: implement it!

(I also set the InputHandler script of Paddle1 back to mouse control.)